feat(python_sdk): create client from connection string
